Extract before and after test to another file. Issue: 41
