+++ tests/test_1build.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 
-import pytest
 import imp
 
 build = imp.load_source('1build', '1build')
@@ -75,20 +74,3 @@ def test_should_print_command_not_found_if_no_command_found_with_given_name(caps
 
     assert invalid_file_error_message in captured.out
 
-
-def test_build_successful_with_before_and_after_command(capsys):
-    build.run("tests/data/build_file_with_before_and_after.yaml", ['file_name', 'build'])
-    captured = capsys.readouterr()
-    expected_cmd_message = "" + dash + "\n" \
-                                       "Name: build\n" \
-                                       "Command: echo 'Running build'\n" \
-                                       "Before: echo 'Running some setup script'\n" \
-                                       "After: echo 'Running some cleanup script'\n" \
-                           + dash + "\n"
-    expected_before_command_output = "Running some setup script"
-    expected_after_command_output = "Running some cleanup script"
-    expected_command_output = "Running build"
-    assert expected_cmd_message in captured.out
-    assert expected_before_command_output in captured.out
-    assert expected_after_command_output in captured.out
-    assert expected_command_output in captured.out
